Refactoring only. No functional change
