+++ server/src/main/java/org/tctalent/server/util/help/HelpLinkHelper.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-package org.tctalent.server.util;
+package org.tctalent.server.util.help;
 
 import java.util.ArrayList;
 import java.util.List;
